Update Maps for Wikia
Using a fake PR to ask a question since issues are disabled here.

What are the plans and timeline for updating the Maps extension on Wikia? It pains me to see such an old version being used, as a ton of fixes have been made and a bunch of new features where added. I'd be happy to help with updating it in return for seeing a modern version of my extension being deployed.

@mszabo-wikia @rybmat